requirements.txt
joblib==1.4.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
scikit-learn==1.4.2
scipy==1.13.0
threadpoolctl==3.4.0
